refactor(grammar): Moved symbol and primeros/siguientes computation to useGrammar hook
